requirements.txt
flask>=3.0.0
discord.py>=2.3.0
google-genai>=1.0.0
certifi>=2024.0.0
aiohttp>=3.9.0
